refactor: join tvOS and iOS AVFoundationPlayback tests
